refactor: create package for kops related functions
